Make the marching-cubes surface work
Pressing c, or passing -s, produced nothing at all. Four separate defects, each
hiding the next.

Nine of a cube's twelve edges never got a vertex. autogenerated.h only ever
assigns vn_id[pxpy], vn_id[pxpz] and vn_id[pypz], the three edges at the +++
corner; the other nine belong to neighbouring cubes and the code meant to copy
them across never made it out of the commented block in fluid.h into the
generated switch. The guard in get_surfacegrid then dropped every triangle that
referenced one of the nine, which is nearly all of them, and printed a line
about it per triangle.

The same cube was triangulated up to eight times. create_cubes() makes, for
every occupied cell, the eight cubes that have it as a corner, so each copy
knows about one corner and none knows about all of them. mergeduplicatecubes()
now folds the copies together after the sort, which is what finally gives each
cube its eight corners.

Every grow of the node list lost its newest node. Nodes live at 1..count, which
is count+1 entries, and resize_node_list was asked to copy count of them. The
lost node read back zeroed - a cell at the origin holding no particles - and
dragged a spike of surface to (0,0,0). Cells also counted particles in an
unsigned char, and cells hold more than 255.

OpenGL read every triangle index as two. The indices were unsigned long, four
bytes on the Windows this was written on and eight here, handed to
glDrawElements as GL_UNSIGNED_INT. Even with the geometry right, that alone
painted garbage over the whole window.

make_vertices_for_cubes() replaces the generated stage: one vertex per crossed
edge, all twelve, placed on the filled cell and stepped towards the empty one by
as much of a cell as its particles fill - the 2005 rule, which used to be a
200-entry table of min(cbrt(n/40),0.99). Two cubes sharing an edge compute
bit-identical coordinates for it, because they compute it from the same cell, so
a map welds them into one indexed mesh, and normals are accumulated per face so
the shading is smooth.

Verified over 600 consecutive frames of the default scene: every vertex inside
the particles' bounding box, every index in range, and triangles to vertices at
2:1, which is the signature of a closed triangulated surface. The default scene
gives about 750 triangles for a compact 6 000-particle drop and 27 000 once it
has broken up into single cells, at 100 to 600 frames a second.

The old generator and autogenerated.h stay in the tree, unused. They are what
the thesis describes.

Also drops the per-frame "Maxparticles in Cell" debug line.

Opus 5 helped authoring this commit.

+++ main.cpp
@@ -95,7 +95,7 @@ void read_24bitbmp ( const char * filename, unsigned char ** pixelinfo, unsigned
 float ** vertex_array_pointer, ** normal_array_pointer;
 unsigned long vn_arraylength, vn_count;
 
-unsigned long ** index_array_pointer;
+unsigned int ** index_array_pointer;
 unsigned long index_arraylength;
 
 float ** particle_coords_pointer, ** particle_colors_pointer;
@@ -205,8 +205,8 @@ int main ( int argc,char** argv )
 	*normal_array_pointer = new float[100];
 	vn_arraylength=10;
 
-	index_array_pointer=new unsigned long*;
-	*index_array_pointer = new unsigned long[100];
+	index_array_pointer=new unsigned int*;
+	*index_array_pointer = new unsigned int[100];
 	index_arraylength=10;
 
 	particle_coords_pointer=new float*;
@@ -569,7 +569,7 @@ void PrintVolumeOfClosedSurface()
 {
 	unsigned long i;
 	float * va;
-	unsigned long * ia;
+	unsigned int * ia;
 	float volume,dv;
 	va = ( *vertex_array_pointer );
 	ia = ( *index_array_pointer );
